// src/siat/catalogos/enums/parametricas.enum.ts
export enum ParametricasEnum {
  EventosSignificativos = 'sincronizarParametricaEventosSignificativos',
  MotivoAnulacion = 'sincronizarParametricaMotivoAnulacion',
  PaisOrigen = 'sincronizarParametricaPaisOrigen',
  TipoDocumentoIdentidad = 'sincronizarParametricaTipoDocumentoIdentidad',
  TipoDocumentoSector = 'sincronizarParametricaTipoDocumentoSector',
  TipoEmision = 'sincronizarParametricaTipoEmision',
  TipoHabitacion = 'sincronizarParametricaTipoHabitacion',
  TipoMetodoPago = 'sincronizarParametricaTipoMetodoPago',
  TipoMoneda = 'sincronizarParametricaTipoMoneda',
  TipoPuntoVenta = 'sincronizarParametricaTipoPuntoVenta',
  TiposFactura = 'sincronizarParametricaTiposFactura',
  UnidadMedida = 'sincronizarParametricaUnidadMedida',
}